@@ -7,14 +7,14 @@ apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.briot.tmmlmss.implementor"
         minSdkVersion 21
         targetSdkVersion 27
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         vectorDrawables.useSupportLibrary = true
         multiDexEnabled true
     }
@@ -22,11 +22,13 @@ android {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-            buildConfigField "String", "HOSTNAME", "\"https://briot-asset-server.herokuapp.com/api/\""
+            //buildConfigField "String", "HOSTNAME", "\"https://briot-asset-server.herokuapp.com/api/\""
+            buildConfigField "String", "HOSTNAME", "\"https://tata-marcopolo-iiot-poc-server.herokuapp.com/\""
         }
 
         debug {
-            buildConfigField "String", "HOSTNAME", "\"https://briot-asset-server.herokuapp.com/api/\""
+            //buildConfigField "String", "HOSTNAME", "\"https://briot-asset-server.herokuapp.com/api/\""
+            buildConfigField "String", "HOSTNAME", "\"https://tata-marcopolo-iiot-poc-server.herokuapp.com/\""
 //            buildConfigField "String", "HOSTNAME", "\"http://10.0.2.2:3000/api/\""
         }
     }
@@ -39,13 +41,14 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.0'
-    implementation 'android.arch.lifecycle:extensions:1.1.1'
-    implementation 'com.android.support:support-v4:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.0.0'
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test:runner:1.1.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0'
     implementation 'net.sourceforge.jtds:jtds:1.3.1'
     implementation group: 'mysql', name: 'mysql-connector-java', version: '8.0.11'
 
@@ -56,15 +59,15 @@ dependencies {
     implementation 'com.squareup.retrofit2:converter-gson:2.4.0'
     implementation 'com.squareup.retrofit2:adapter-rxjava2:2.4.0'
 
-    implementation "android.arch.navigation:navigation-fragment:1.0.0-alpha01"
-    implementation "android.arch.navigation:navigation-ui:1.0.0-alpha01"
+    implementation 'androidx.navigation:navigation-fragment:2.0.0-rc02'
+    implementation 'androidx.navigation:navigation-ui:2.0.0-rc02'
 
     implementation 'de.hdodenhof:circleimageview:2.2.0'
     implementation 'com.github.Pierry:Progress:v1.1'
     implementation 'com.pascalwelsch.arrayadapter:arrayadapter:1.3.0'
 
-    implementation 'com.android.support:multidex:1.0.3'
-    implementation 'com.android.support:recyclerview-v7:27.1.1'
+    implementation 'androidx.multidex:multidex:2.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
 }
 
 kapt {
