@@ -1,17 +1,16 @@
 package com.briot.tmmlmss.implementor.ui.main
 
-import android.arch.lifecycle.Observer
-import android.arch.lifecycle.ViewModelProviders
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProviders
 import android.content.Context
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v7.app.AppCompatActivity
+import androidx.fragment.app.Fragment
+import androidx.appcompat.app.AppCompatActivity
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.Navigation
 import com.briot.tmmlmss.implementor.MainActivity
-
 import com.briot.tmmlmss.implementor.R
 import com.briot.tmmlmss.implementor.repository.local.PrefConstants
 import com.briot.tmmlmss.implementor.repository.local.PrefRepository
@@ -21,7 +20,7 @@ import kotlinx.android.synthetic.main.login_fragment.*
 import android.view.inputmethod.InputMethodManager
 
 
-class LoginFragment : Fragment() {
+class LoginFragment : androidx.fragment.app.Fragment() {
 
     companion object {
         fun newInstance() = LoginFragment()
@@ -40,13 +39,16 @@ class LoginFragment : Fragment() {
         viewModel = ViewModelProviders.of(this).get(LoginViewModel::class.java)
         // TODO: Use the ViewModel
 
+        username.requestFocus()
+        password.requestFocus()
+
         viewModel.user.observe(this, Observer<User> {
             MainActivity.hideProgress(this.progress)
             this.progress = null
 
             if (it != null && it != viewModel.invalidUser && it.token != null) {
                 PrefRepository.singleInstance.setKeyValue(PrefConstants().USERLOGGEDIN, username.text.toString())
-                PrefRepository.singleInstance.setKeyValue(PrefConstants().USER_TOKEN,  it.token!!)
+                PrefRepository.singleInstance.setKeyValue(PrefConstants().USER_TOKEN, it.token!!)
                 Navigation.findNavController(login).navigate(R.id.action_loginFragment_to_homeFragment)
                 this.context?.let { it1 -> PrefRepository.singleInstance.serializePrefs(it1) }
             } else {
@@ -65,7 +67,7 @@ class LoginFragment : Fragment() {
 
         login.setOnClickListener {
             val keyboard = activity!!.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
-            keyboard.hideSoftInputFromWindow(activity!!.currentFocus.getWindowToken(), 0)
+            keyboard.hideSoftInputFromWindow(activity?.currentFocus?.getWindowToken(), 0)
 
             this.progress = MainActivity.showProgressIndicator(this.activity as AppCompatActivity, "Please wait")
             viewModel.loginUser(username.text.toString(), password.text.toString())
