@@ -1,8 +1,8 @@
 package com.briot.tmmlmss.implementor.ui.main
 
-import android.arch.lifecycle.ViewModelProviders
+import androidx.lifecycle.ViewModelProviders
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -17,7 +17,7 @@ import kotlinx.android.synthetic.main.login_fragment.*
 import kotlinx.android.synthetic.main.main_fragment.*
 import java.util.concurrent.TimeUnit
 
-class MainFragment : Fragment() {
+class MainFragment : androidx.fragment.app.Fragment() {
 
     companion object {
         fun newInstance() = MainFragment()
@@ -41,11 +41,11 @@ class MainFragment : Fragment() {
         if (userToken.isNotEmpty()) {
             Observable.timer(2000, TimeUnit.MILLISECONDS)
                     .observeOn(AndroidSchedulers.mainThread())
-                    .subscribe( { Navigation.findNavController(img).navigate(R.id.homeFragment) } );
+                    .subscribe({ Navigation.findNavController(img).navigate(R.id.homeFragment) });
         } else {
             Observable.timer(2000, TimeUnit.MILLISECONDS)
                     .observeOn(AndroidSchedulers.mainThread())
-                    .subscribe( { Navigation.findNavController(img).navigate(R.id.action_mainFragment_to_loginFragment) } );
+                    .subscribe({ Navigation.findNavController(img).navigate(R.id.action_mainFragment_to_loginFragment) });
         }
 
 
