@@ -51,7 +51,7 @@ class PrefConstants {
     public val PUTAWAYLOCATION = "PUTAWAYLOCATION"
     public val PICKLISTID = "PICKLISTID"
     public val USER_TOKEN = "USERTOKEN"
-    public val SELECTED_AUDIT_ID = "SELECTED_AUDIT_ID"
+    public val SELECTED_MACHINE_ID = "SELECTED_MACHINE_ID"
     public val SELECTED_AUDIT_SITEID = "SELECTED_AUDIT_SITEID"
     public val SELECTED_AUDIT_LOCATIONID = "SELECTED_AUDIT_LOCATIONID"
     public val SELECTED_AUDIT_SUBLOCATIONID = "SELECTED_AUDIT_SUBLOCATIONID"
