@@ -2,6 +2,6 @@ package com.briot.tmmlmss.implementor.ui.main
 
 import androidx.lifecycle.ViewModel;
 
-class HomeViewModel : ViewModel() {
+class PendingItemDashboardViewModel : ViewModel() {
     // TODO: Implement the ViewModel
 }
